+++ src/docs/previews/slider/vertical/tailwind/index.svelte
@@ -2,7 +2,7 @@
 	import { createSlider, melt } from '$lib/index.js';
 
 	const {
-		elements: { root, range, thumb },
+		elements: { root, range, thumbs },
 	} = createSlider({
 		defaultValue: [30],
 		max: 100,
@@ -17,8 +17,9 @@
 	<span class="h-[200px] w-full bg-black/40">
 		<span use:melt={$range} class="w-full bg-white" />
 	</span>
+
 	<span
-		use:melt={$thumb()}
+		use:melt={$thumbs[0]}
 		class="h-5 w-5 rounded-full bg-white focus:ring-4 focus:!ring-black/40"
 	/>
 </span>