+++ src/routes/(landing-ui)/slider.svelte
@@ -3,21 +3,22 @@
 	import { createSlider, melt } from '$lib/index.js';
 
 	const {
-		elements: { root, range, thumb },
+		elements: { root, range, thumbs },
 	} = createSlider({
 		defaultValue: [60],
 		max: 100,
 	});
+
 	let className = '';
 	export { className as class };
 </script>
 
-<span use:melt={$root} class={cn('relative flex  w-[15rem] items-center', className)}>
+<span use:melt={$root} class={cn('relative flex w-[15rem] items-center', className)}>
 	<span class="block h-1.5 w-full rounded-full bg-neutral-400 dark:bg-neutral-700">
 		<span use:melt={$range} class="h-1.5 rounded-full bg-magnum-400" />
 	</span>
 	<span
-		use:melt={$thumb()}
+		use:melt={$thumbs[0]}
 		class="block rounded-full bg-white shadow square-6 focus:ring-4 focus:ring-magnum-600 dark:bg-white dark:shadow-none"
 	/>
 </span>